improved tests on materialeditor

+++ src/materialeditor.h
@@ -12,7 +12,6 @@
 #include <OgreTechnique.h>
 #include <OgrePass.h>
 
-
 namespace Ui {
 class MaterialEditor;
 }
@@ -21,18 +20,18 @@ class MaterialHighlighter;
 
 class MaterialEditor : public QDialog
 {
-        Q_OBJECT
-        
-    public:
-        explicit MaterialEditor(QWidget *parent = 0);
-        virtual ~MaterialEditor();
-        void setMaterialText(const QString &_mat);
-        void setMaterial(const QString& _material);
-        std::string getMaterialText() const;
-        std::string getMaterialName() const;
-        bool isScrollAreaEnabled() const;
-        void setTechFields(const QMap<int, Ogre::Pass *> &_techMap, const QList<QString> &_passList);
-        Ui::MaterialEditor* getUI() const;
+    Q_OBJECT
+    
+public:
+    explicit MaterialEditor(QWidget *parent = 0);
+    virtual ~MaterialEditor();
+    void setMaterialText(const QString &_mat);
+    void setMaterial(const QString& _material);
+    std::string getMaterialText() const;
+    std::string getMaterialName() const;
+    bool isScrollAreaEnabled() const;
+    void setTechFields(const QMap<int, Ogre::Pass *> &_techMap, const QList<QString> &_passList);
+    Ui::MaterialEditor* getUI() const;
 
 private slots:
     void on_buttonEditAmbientColor_clicked();
@@ -67,11 +66,11 @@ private slots:
 
     void on_checkBoxUseVertexColorToEmissive_toggled(bool checked);
 
-    void on_alphaDifuse_valueChanged(double arg1);
+    void on_alphaDifuse_valueChanged(float arg1);
 
-    void on_alphaSpecular_valueChanged(double arg1);
+    void on_alphaSpecular_valueChanged(float arg1);
 
-    void on_shineSpecular_valueChanged(double arg1);
+    void on_shineSpecular_valueChanged(float arg1);
 
     void on_newTechnique_clicked();
